Improvements for the `DiffCalculation` plugin
Removes duplicated input port definitions and gets rid of class
variables used to set the defaults for metadata inputs. Since these are
set on the process spec, they should be retrieved from the process spec
or the actual inputs, depending on the use case. For example in the
`prepare_for_submission` the actual values passed to the process should
be used, otherwise making them configurable options does not make sense
if we are always going to override them with the class variable.